linoleum: Add MaudeLotrImageGenSafety example, and fix thread safe class
Add a global lock on calls to the Maude intepreter to prevent SIGSEGV
crashes due to concurrent term parsing or rewrite. The current crude
lock should be revised in further commits

The property is not working as expected as currently the rewrite
rules and truth value calculation are not working as expected. As
the unit test MaudeLotrImageGenSafetyTest this needs further investigation